+++ provisioner/resource_provisioner.go
@@ -1,4 +1,4 @@
-package main
+package provisioner
 
 import (
 	"context"
